+++ internal/types/definitions.go
@@ -2,6 +2,7 @@ package types
 
 import (
 	"net/http"
+	"sync"
 
 	client "github.com/pingidentity/pingfederate-go-client/v1210/configurationapi"
 	"github.com/pingidentity/terraform-provider-pingfederate/internal/version"
@@ -9,16 +10,17 @@ import (
 
 // Configuration used by the provider and resources
 type ProviderConfiguration struct {
-	HttpsHost      string
-	Transport      *http.Transport
-	Username       *string
-	Password       *string
-	AccessToken    *string
-	TokenUrl       *string
-	ClientId       *string
-	ClientSecret   *string
-	Scopes         []string
-	ProductVersion version.SupportedVersion
+	HttpsHost          string
+	Transport          *http.Transport
+	Username           *string
+	Password           *string
+	AccessToken        *string
+	TokenUrl           *string
+	ClientId           *string
+	ClientSecret       *string
+	Scopes             []string
+	ProductVersion     version.SupportedVersion
+	KeypairCreateMutex *sync.Mutex
 }
 
 // Configuration passed to resources